services/ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
openai==1.3.8
anthropic==0.7.8
sqlalchemy==2.0.23
asyncpg==0.29.0
redis==5.0.1
nats-py==2.6.0
httpx==0.25.2
python-dotenv==1.0.0
python-multipart==0.0.6
Pillow==10.1.0
requests==2.31.0
aiofiles==23.2.1
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
celery==5.3.4
qdrant-client==1.7.0